Move test utils back to original package
