+++ deferred-resources/src/androidTest/java/com/backbase/deferredresources/DeferredBooleanTest.kt
@@ -2,7 +2,6 @@ package com.backbase.deferredresources
 
 import android.os.Build
 import com.backbase.deferredresources.test.R
-import com.backbase.deferredresources.test.context
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
