+++ deferred-resources/src/androidTest/java/com/backbase/deferredresources/DeferredDimensionTest.kt
@@ -1,7 +1,6 @@
 package com.backbase.deferredresources
 
 import com.backbase.deferredresources.test.R
-import com.backbase.deferredresources.test.context
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
