+++ deferred-resources/src/androidTest/java/com/backbase/deferredresources/DeferredTextTest.kt
@@ -4,7 +4,6 @@ import android.graphics.Typeface
 import android.text.SpannedString
 import android.text.style.StyleSpan
 import com.backbase.deferredresources.test.R
-import com.backbase.deferredresources.test.context
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
